Partners: Update with 2026 partners
Also reorganizes the data to set up a future script
