+++ partner-data/community-groups.json
@@ -1,0 +1,47 @@
+[
+   {
+      "name": "Wash Park Crewzers",
+      "logo": "/partners/wash-park-crewzers.png",
+      "instagram": "@wash_park_crewzers",
+      "ready": true
+   },
+   {
+      "name": "SkidMarks Bike Club",
+      "logo": "/partners/skidmarks.png",
+      "instagram": "@skidmarksbc",
+      "ready": false
+   },
+   {
+      "name": "Littleton Social Cycle",
+      "logo": "/partners/littleton-social-cycle.png",
+      "url": "https://www.littletonsocialcycle.com/",
+      "instagram": "@littletonsocialcycle",
+      "ready": true
+   },
+   {
+      "name": "Chains of Strength",
+      "logo": "/partners/chains-of-strength.png",
+      "instagram": "@chains.of.strength",
+      "ready": false
+   },
+   {
+      "name": "Bike and Brew Denver",
+      "logo": "/partners/bike-and-brew.png",
+      "instagram": "@bikeandbrewdenver",
+      "ready": true
+   },
+   {
+      "name": "Denver Bicycle Lobby",
+      "logo": "/partners/denver-bicycle-lobby.png",
+      "url": "https://denverbicyclelobby.com/",
+      "instagram": "@denverbicyclelobby",
+      "ready": true
+   },
+   {
+      "name": "Denver Critical Mass",
+      "logo": "/partners/denver-critical-mass.png",
+      "url": "https://denvercriticalmass.org/",
+      "instagram": "@criticalmassdenver",
+      "ready": false
+   }
+]
